% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{coef_plot}
\alias{coef_plot}
\title{Coefficient plot of fixed effects (and sigma^2) with confidence intervals}
\usage{
coef_plot(object, level = 0.95, intercept = FALSE)
}
\arguments{
\item{object}{a fitted \code{"SDALGCP2"} object.}

\item{level}{confidence level.}

\item{intercept}{logical; include the intercept.}
}
\value{
a \code{ggplot} object.
}
\description{
Coefficient plot of fixed effects (and sigma^2) with confidence intervals
}
\examples{
\donttest{
data(sdalgcp_data)
fit <- sdalgcp(cases ~ x1 + offset(log(pop)), data = sdalgcp_data,
               control = sdalgcp_control(n_sim = 2000, burnin = 500, thin = 5,
                                         reanchor = 0))
coef_plot(fit)
}
}
