% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.SDALGCP2}
\alias{confint.SDALGCP2}
\title{Wald confidence intervals for an SDALGCP2 fit}
\usage{
\method{confint}{SDALGCP2}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class \code{"SDALGCP2"}.}

\item{parm}{parameters to report (names or indices); default all.}

\item{level}{confidence level.}

\item{...}{unused.}
}
\value{
a matrix of lower/upper confidence limits.
}
\description{
Wald confidence intervals for an SDALGCP2 fit
}
\examples{
\donttest{
data(sdalgcp_data)
fit <- sdalgcp(cases ~ x1 + offset(log(pop)), data = sdalgcp_data,
               control = sdalgcp_control(n_sim = 2000, burnin = 500, thin = 5,
                                         reanchor = 0))
confint(fit)
}
}
