% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot.SDALGCP2}
\alias{plot.SDALGCP2}
\title{Plot an SDALGCP2 fit (the phi profile deviance)}
\usage{
\method{plot}{SDALGCP2}(x, ...)
}
\arguments{
\item{x}{an \code{"SDALGCP2"} object.}

\item{...}{passed to \code{\link{phi_profile}}.}
}
\value{
invisibly, the profile (see \code{\link{phi_profile}}).
}
\description{
Plot an SDALGCP2 fit (the phi profile deviance)
}
