% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.SDALGCP2}
\alias{print.SDALGCP2}
\title{Print an SDALGCP2 fit}
\usage{
\method{print}{SDALGCP2}(x, ...)
}
\arguments{
\item{x}{an \code{"SDALGCP2"} object.}

\item{...}{unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an SDALGCP2 fit
}
