% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.SDALGCP2}
\alias{print.summary.SDALGCP2}
\title{Print a summary of an SDALGCP2 fit}
\usage{
\method{print}{summary.SDALGCP2}(x, ...)
}
\arguments{
\item{x}{a \code{"summary.SDALGCP2"} object.}

\item{...}{unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a summary of an SDALGCP2 fit
}
