% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDALGCP2-package.R
\docType{package}
\name{SDALGCP2-package}
\alias{SDALGCP2-package}
\title{SDALGCP2: Fast Spatially Discrete Approximation to Log-Gaussian Cox Processes}
\description{
A faster, modernised successor to the \pkg{SDALGCP} package for fitting a
spatially discrete approximation to a log-Gaussian Cox process (SDA-LGCP) to
spatially aggregated disease counts. Performance-critical steps are
implemented in C++ via \pkg{RcppArmadillo}. See \code{DESIGN.md} in the
package source for the full design and statistical roadmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/olatunjijohnson/SDALGCP2}
  \item \url{https://olatunjijohnson.github.io/SDALGCP2/}
  \item Report bugs at \url{https://github.com/olatunjijohnson/SDALGCP2/issues}
}

}
\author{
\strong{Maintainer}: Olatunji Johnson \email{teejaystatistics@gmail.com}

Authors:
\itemize{
  \item Emanuele Giorgi
  \item Peter Diggle
}

}
\keyword{internal}
