% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.SDALGCP2}
\alias{summary.SDALGCP2}
\title{Summary of an SDALGCP2 fit}
\usage{
\method{summary}{SDALGCP2}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"SDALGCP2"} from \code{\link{mcml_fit}}.}

\item{...}{unused.}
}
\value{
an object of class \code{"summary.SDALGCP2"} with a coefficient table.
}
\description{
Summary of an SDALGCP2 fit
}
