% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{coef_plot}
\alias{coef_plot}
\title{Coefficient plot of fixed effects (and sigma^2) with confidence intervals}
\usage{
coef_plot(object, level = 0.95, intercept = FALSE)
}
\arguments{
\item{object}{a fitted \code{"SDALGCP2"} object.}

\item{level}{confidence level.}

\item{intercept}{logical; include the intercept.}
}
\value{
a \code{ggplot} object.
}
\description{
Coefficient plot of fixed effects (and sigma^2) with confidence intervals
}
