% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.SDALGCP2}
\alias{confint.SDALGCP2}
\title{Wald confidence intervals for an SDALGCP2 fit}
\usage{
\method{confint}{SDALGCP2}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class \code{"SDALGCP2"}.}

\item{parm}{parameters to report (names or indices); default all.}

\item{level}{confidence level.}

\item{...}{unused.}
}
\value{
a matrix of lower/upper confidence limits.
}
\description{
Wald confidence intervals for an SDALGCP2 fit
}
