% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot.SDALGCP2}
\alias{plot.SDALGCP2}
\title{Plot an SDALGCP2 fit (the phi profile deviance)}
\usage{
\method{plot}{SDALGCP2}(x, ...)
}
\arguments{
\item{x}{an \code{"SDALGCP2"} object.}

\item{...}{passed to \code{\link{phi_profile}}.}
}
\value{
invisibly, the profile (see \code{\link{phi_profile}}).
}
\description{
Plot an SDALGCP2 fit (the phi profile deviance)
}
\examples{
\donttest{
data(sdalgcp_data)
fit <- SDALGCP2(cases ~ x1 + offset(log(pop)),
                sf::st_drop_geometry(sdalgcp_data), sdalgcp_data, delta = 1.2,
                control.mcmc = control_mcmc(n.sim = 2000, burnin = 500, thin = 5))
plot(fit)   # profile deviance for the spatial scale phi
}
}
